refactor(DCMAW-8413): move types to their own folder
