+++ src/credentialOfferViewer/types/CredentialOfferResponse.ts
