+++ src/dbsDocumentBuilder/types/DbsInputData.ts
@@ -1,0 +1,18 @@
+export interface DbsInputData {
+    "issuance-day"?: string;
+    "issuance-month"?: string;
+    "issuance-year"?: string;
+    "birth-day"?: string;
+    "birth-month"?: string;
+    "birth-year"?: string;
+    firstName?: string;
+    lastName?: string;
+    subBuildingName?: string;
+    buildingName?: string;
+    streetName?: string;
+    addressLocality?: string;
+    addressCountry?: string;
+    postalCode?: string;
+    certificateNumber?: string;
+    applicationNumber?: string;
+}
