+++ src/ninoDocumentBuilder/types/NinoCredentialSubject.ts
@@ -1,0 +1,9 @@
+import {CredentialSubject} from "../../types/CredentialSubject";
+
+export interface NinoCredentialSubject extends CredentialSubject {
+    socialSecurityRecord?: SocialSecurityRecord[];
+}
+
+export interface SocialSecurityRecord {
+    personalNumber: string;
+}
