+++ src/ninoDocumentBuilder/types/NinoInputData.ts
@@ -1,0 +1,6 @@
+export interface NinoInputData {
+  title?: string;
+  givenName?: string;
+  familyName?: string;
+  nino?: string;
+}
