+++ src/stsStubAccessToken/types/AccessTokenPayload.ts
