+++ src/stsStubAccessToken/types/PreAuthorizedCodePayload.ts
